requirements.txt
python-telegram-bot==13.0
geopy==2.0.0
requests==2.24.0
